refactor(acl): split ManageManualMemberController into Add/RemoveManualMemberController
ManageManualMemberController had two public methods (add/remove),
violating the single-action controller convention.

- AddManualMemberController — invokable, POST acl.member.add
- RemoveManualMemberController — invokable, DELETE acl.member.remove
- Delete ManageManualMemberController
- Update routes
- Rename test to ManualMemberControllerTest, update description strings
